update the params of the service to use the new getcontactlist model, added total count to the response

+++ BackEnd/ContactsAPI/ContactsAPI/Controllers/ContactController.cs
@@ -4,7 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using ContactsAPI.Models;
 using ContactsAPI.Services;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.FileProviders;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -21,40 +24,56 @@ namespace ContactsAPI.Controllers
         }
         // GET: api/<controller>
         [HttpGet]
-        public async Task<PaginatedList<Contact>> Get()
+        public async Task<Ok<PaginatedList<Contact>>> Get(GetContactList model)
         {
-            return await service.GetContacts();
+            return TypedResults.Ok(await service.GetContacts(model));
         }
 
         // GET api/<controller>/5
         [HttpGet("{contactId}")]
-        public async Task<Contact> Get(int contactId)
+        public async Task<Results<Ok<Contact>, NotFound>> Get(int contactId)
         {
-            return await service.GetContactById(contactId);
+            var contact = await service.GetContactById(contactId);
+
+            if (contact == null) return TypedResults.NotFound();
+
+            return TypedResults.Ok(contact);
         }
 
         // POST api/<controller>
         [HttpPost]
-        public async Task<IActionResult> Post(CreateContact value)
+        public async Task<Created<Contact>> Post(CreateContact value)
         {
             var contact = await service.CreateContact(value);
-            return Created($"/api/contact/{contact.Id}", contact);
+
+            return TypedResults.Created($"/api/contact/{contact.Id}", contact);
         }
 
         // PUT api/<controller>/5
         [HttpPut("{contactId}")]
-        public async Task<IActionResult> Put(int contactId, UpdateContact value)
+        public async Task<Results<Ok<Contact>, NotFound>> Put(int contactId, UpdateContact value)
         {
+            if (await service.GetContactById(contactId) == null)
+            {
+                return TypedResults.NotFound();
+            }
+
             var contact = await service.UpdateContact(contactId, value);
-            return Ok(contact);
+
+            return TypedResults.Ok(contact);
         }
 
         // DELETE api/<controller>/5
         [HttpDelete("{contactId}")]
-        public async Task<IActionResult> Delete(int contactId)
+        public async Task<Results<NoContent, NotFound>> Delete(int contactId)
         {
+            if (await service.GetContactById(contactId) == null)
+            {
+                return TypedResults.NotFound();
+            }
+
             await service.DeleteContact(contactId);
-            return NoContent();
+            return TypedResults.NoContent();
         }
     }
 }
